backend/ml-services/url-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
pymongo==4.6.1

# ML Libraries
torch==2.1.0
torch-geometric==2.4.0
networkx==3.2.1
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2

# Web & Network
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0
dnspython==2.4.2
python-whois==0.8.0
cryptography==41.0.7

# Utilities
tldextract==5.0.1
urllib3==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
